Changes callbacks to method calls in controller actions.
This is a personal preference, but one that I believe many other developers are
following too. Callbacks on models can hide the complexity that is happening in
a controller action. Making the calls explicitly within the action shows exactly
what that action is up to. Other things this helps with are tests or admin
actions, where you don't get unintended side effects if you create an object
outside of its normal life. There are some blog posts on this, for example
[here](http://adequate.io/culling-the-activerecord-lifecycle) and [here](https://www.reinteractive.net/posts/130-activerecord-callbacks-considered-harmful).